+++ Alignment/Alignment/AlignTrTools/src/AlignTrackMonitor.cpp
@@ -1,4 +1,4 @@
-// $Id: AlignTrackMonitor.cpp,v 1.2 2007-04-25 14:07:06 jblouw Exp $
+// $Id: AlignTrackMonitor.cpp,v 1.3 2007-05-14 17:47:31 cattanem Exp $
 //
 
 //-----------------------------------------------------------------------------
@@ -540,7 +540,7 @@ AlignTrackMonitor::fillVariables ( const LHCb::Track* aTrack,
         //**********************************************************************
         // Two hits in same "sub-layer" of same module
         if ( (theOTID.uniqueModule() == theOTID2.uniqueModule()) &&
-             (abs(theOTID.straw()-theOTID2.straw())<5) ) {
+             (abs(int(theOTID.straw()-theOTID2.straw()))<5) ) {
           if ( msgLevel( MSG::DEBUG ) )
             debug() << "Found OT double hit" << endmsg;
           ++m_nOTDoubleHits;
@@ -769,7 +769,7 @@ bool AlignTrackMonitor::isNeighbouringHit ( LHCb::STChannelID clusID,
 
   // Same ladder within 2 strips aside
   if ( (hitID.uniqueSector() == clusID.uniqueSector()) &&
-       (abs(hitStrip-clusID.strip())<=m_nStripsTol) ) return true;
+       (abs(int(hitStrip-clusID.strip()))<=m_nStripsTol) ) return true;
   
   return false;
 }
@@ -788,11 +788,11 @@ bool AlignTrackMonitor::isNeighbouringHit ( LHCb::OTChannelID timeID,
   
   // In module 9 of quarters 0 and 2, only 2*32 straws
   if ( !(timeID.quarter()%2) && (int(timeID.module()) == m_nModules) ) {
-    if ( (abs((hitStraw%int(m_nStraws/4))
-              -(timeID.straw()%int(m_nStraws/4)))<=m_nStrawsTol) ) return true;
+    if ( (abs(int((hitStraw%int(m_nStraws/4))
+              -(timeID.straw()%int(m_nStraws/4))))<=m_nStrawsTol) ) return true;
   }
-  else if ( (abs((hitStraw%int(m_nStraws/2))
-                 -(timeID.straw()%int(m_nStraws/2)))<=m_nStrawsTol) ) return true;
+  else if ( (abs(int((hitStraw%int(m_nStraws/2))
+                 -(timeID.straw()%int(m_nStraws/2))))<=m_nStrawsTol) ) return true;
 
   return false;
 }
